requirements.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.2
future==0.18.2
httplib2==0.17.0
httpretty==1.0.2
idna==2.9
lxml==4.5.0
oauth2==1.9.0.post1
oauthlib==3.1.0
pycparser==2.20
python-telegram-bot==12.4.2
PyTumblr==0.1.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tornado==6.0.4
urllib3==1.25.8
